New folder (2)/requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
celery==5.3.4
requests==2.31.0
lxml==4.9.3
pandas==2.1.3
webdriver-manager==4.0.1
python-dotenv==1.0.0
schedule==1.2.0
